add affirmation text to recyclerview
