+++ .idea/misc.xml
@@ -3,8 +3,10 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/activity_main.xml" value="0.246875" />
+        <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/fragment_welcome.xml" value="0.246875" />
         <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/onboarding1_layout_resource_file.xml" value="0.246875" />
-        <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/onboarding1_list_layout.xml" value="0.1" />
+        <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/onboarding1_list_layout.xml" value="0.268" />
         <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/onboarding2_layout_resource_file.xml" value="0.246875" />
         <entry key="..\:/Users/ayana/AndroidStudioProjects/PregnancyApp/app/src/main/res/layout/onboarding_layout.xml" value="0.246875" />
         <entry key="..\:/Users/keyar/AndroidStudioProjects/app/src/main/res/layout/activity_main.xml" value="0.22333333333333336" />
