+++ app/src/main/java/com/example/pregnancy_app/adapter/OnboardRecyclerViewAdapter.kt
@@ -1,5 +1,6 @@
 package com.example.pregnancy_app.adapter
 
+import android.location.GnssAntennaInfo
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -8,12 +9,17 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.pregnancy_app.R
 
-class OnboardRecyclerViewAdapter : RecyclerView.Adapter<OnboardRecyclerViewAdapter.OnboardViewHolder>() {
-    private val affirmationText = mutableListOf<String>()
+class OnboardRecyclerViewAdapter(val affirmationText: MutableList<String>, val listener: Listener) : RecyclerView.Adapter<OnboardRecyclerViewAdapter.OnboardViewHolder>() {
 
     inner class OnboardViewHolder(view: View) : RecyclerView.ViewHolder(view) {
         val imageView: ImageView = view.findViewById(R.id.imageView3)
         val textView: TextView = view.findViewById(R.id.textView3)
+
+        init {
+            imageView.setOnClickListener {
+                listener.onImageClicked(adapterPosition, view)
+            }
+        }
     }
     // Create new views
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OnboardViewHolder {
@@ -28,5 +34,9 @@ class OnboardRecyclerViewAdapter : RecyclerView.Adapter<OnboardRecyclerViewAdapt
     // Return the size of your dataset
     override fun getItemCount(): Int {
         return affirmationText.size
+    }
+
+    interface Listener{
+        fun onImageClicked(index: Int, view: View)
     }
 }
