+++ app/src/main/java/com/example/pregnancy_app/ui/LoginScreen.kt
@@ -1,10 +1,11 @@
-package com.example.pregnancy_app
+package com.example.pregnancy_app.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.pregnancy_app.R
 
 class LoginScreen:Fragment () {
 
@@ -14,6 +15,5 @@ class LoginScreen:Fragment () {
     ): View? {
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.login_screen_fragment, container, false)
-
     }
 }
