+++ app/src/main/java/com/example/pregnancy_app/ui/OnboardingTwo.kt
@@ -1,4 +1,4 @@
-package com.example.pregnancy_app
+package com.example.pregnancy_app.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,25 +7,23 @@ import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
-import androidx.recyclerview.widget.RecyclerView
-import com.example.pregnancy_app.adapter.OnboardRecyclerViewAdapter
+import com.example.pregnancy_app.R
 
-class OnboardingOne: Fragment() {
-    private lateinit var adapter: OnboardRecyclerViewAdapter
+class OnboardingTwo: Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.onboarding1_list_layout, container, false)
+        return inflater.inflate(R.layout.onboarding2_layout_resource_file, container, false)
     }
-
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        adapter= OnboardRecyclerViewAdapter()
-        val recyclerView = view.findViewById<RecyclerView>(R.id.photos_grid)
-        recyclerView.adapter = adapter
-        }
 
+        view.findViewById<Button>(R.id.button_onboarding2).setOnClickListener { view ->
+            view.findNavController().navigate(R.id.action_onboardingTwo_to_loginScreen)
+
+        }
+    }
 }
