+++ app/src/main/java/com/example/pregnancy_app/ui/WelcomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.pregnancy_app
+package com.example.pregnancy_app.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,21 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.navigation.findNavController
-
-
-
+import com.example.pregnancy_app.R
 
 class WelcomeFragment : Fragment() {
-
-
-    /* override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-
-        }
-    }
-
-     */
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -37,8 +25,5 @@ class WelcomeFragment : Fragment() {
         view.findViewById<Button>(R.id.button_activity_main).setOnClickListener { view ->
             view.findNavController().navigate(R.id.action_welcomeFragment_to_onboardingOne)
         }
-
-
-
-                }
-            }
+    }
+}
