+++ app/src/main/res/layout/onboarding_layout.xml
@@ -1,35 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
 <com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
+    android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:padding="2dp"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".OnboardingOne">
+    android:orientation="vertical"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <TextView
         android:id="@+id/textView3"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:padding="2dp"
+        android:padding="4dp"
         android:text="TextView"
-        android:layout_marginStart="45dp"
-        android:layout_marginTop="180dp"/>
+        android:layout_marginTop="155dp"/>
     <!-- Figure out why vertical constraints can't be added to textview-->
 
     <ImageView
         android:id="@+id/imageView3"
         android:layout_width="wrap_content"
-        android:layout_height="170dp"
-        app:layout_constraintBottom_toTopOf="@+id/textView3"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.7"
-        android:padding="4dp"
-        android:focusable="true"
-        android:clickable="true"
+        android:layout_height="155dp"
         android:adjustViewBounds="true"
+        android:clickable="true"
+        android:focusable="true"
+        android:padding="6dp"
+        android:scaleType="fitXY"
         app:srcCompat="@drawable/pregnant_image_3" />
 </com.google.android.material.card.MaterialCardView>
