+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,28 +7,16 @@
 
     <fragment
         android:id="@+id/onboardingOne"
-        android:name="com.example.pregnancy_app.OnboardingOne"
+        android:name="com.example.pregnancy_app.ui.OnboardingOne"
         android:label="OnboardingOne"
         tools:layout="@layout/onboarding1_list_layout">
-        <action
-            android:id="@+id/action_onboardingOne_to_welcomeFragment"
-            app:destination="@id/welcomeFragment" />
         <action
             android:id="@+id/action_onboardingOne_to_onboardingTwo"
             app:destination="@id/onboardingTwo" />
     </fragment>
     <fragment
-        android:id="@+id/welcomeFragment"
-        android:name="com.example.pregnancy_app.WelcomeFragment"
-        android:label="WelcomeFragment"
-        tools:layout="@layout/fragment_welcome">
-        <action
-            android:id="@+id/action_welcomeFragment_to_onboardingOne"
-            app:destination="@id/onboardingOne" />
-    </fragment>
-    <fragment
         android:id="@+id/onboardingTwo"
-        android:name="com.example.pregnancy_app.OnboardingTwo"
+        android:name="com.example.pregnancy_app.ui.OnboardingTwo"
         android:label="OnboardingTwo"
         tools:layout="@layout/onboarding2_layout_resource_file">
         <action
@@ -40,11 +28,20 @@
     </fragment>
     <fragment
         android:id="@+id/loginScreen"
-        android:name="com.example.pregnancy_app.LoginScreen"
+        android:name="com.example.pregnancy_app.ui.LoginScreen"
         android:label="LoginScreen"
         tools:layout="@layout/login_screen_fragment">
         <action
             android:id="@+id/action_loginScreen_to_onboardingTwo"
             app:destination="@id/onboardingTwo" />
+    </fragment>
+    <fragment
+        android:id="@+id/welcomeFragment"
+        android:name="com.example.pregnancy_app.ui.WelcomeFragment"
+        android:label="fragment_welcome"
+        tools:layout="@layout/fragment_welcome" >
+        <action
+            android:id="@+id/action_welcomeFragment_to_onboardingOne"
+            app:destination="@id/onboardingOne" />
     </fragment>
 </navigation>
